bleh

+++ include/g_local.h
@@ -148,7 +148,7 @@ extern int timelimit, fraglimit, teamplay, deathmatch, framecount, coop, skill;
 extern int sv_extensions;
 //extern float	rj;
 
-#define	EDICT_TO_PROG(e) ((byte *)(e) - (byte *)g_edicts)
+#define	EDICT_TO_PROG(e) ((int)((byte *)(e) - (byte *)g_edicts))
 #define PROG_TO_EDICT(e) ((gedict_t *)((byte *)g_edicts + (e)))
 
 void G_Printf(const char *fmt, ...) PRINTF_FUNC(1);
@@ -457,7 +457,7 @@ void visible_to(gedict_t *viewer, gedict_t *first, int len, byte *visible);
 
 // Work around for the fact that QVM dos not support ".*s" in printf() family functions.
 // It retuns dots array filled with dots, amount of dots depends of how long cmd name and longest cmd name.
-char* make_dots(char *dots, size_t dots_len, int cmd_max_len, char *cmd);
+char* make_dots(char *dots, size_t dots_len, size_t cmd_max_len, char *cmd);
 
 //
 // subs.c
